+++ examples/alice_chopper/lib/main.dart
@@ -339,7 +339,7 @@ class _MyAppState extends State<MyApp> {
 
   /// Run all Chopper HTTP requests
   Future<void> _runChopperHttpRequests() async {
-    _alice.tag('chopper-flow');
+    _alice.tag(tag: 'chopper-flow');
     await Future.wait([
       _albumsRequests(),
       _articlesRequests(),