+++ examples/alice_http/lib/main.dart
@@ -66,7 +66,7 @@ class _MyAppState extends State<MyApp> {
       'userId': '1',
     };
 
-    _alice.tag('test-flow');
+    _alice.tag(tag: 'test-flow');
     http
         .post(Uri.https('jsonplaceholder.typicode.com', '/posts'), body: body)
         .interceptWithAlice(_aliceHttpAdapter, body: body);