+++ packages/alice/lib/src/core/alice.dart
@@ -48,4 +48,12 @@ class Alice {
 
   /// Adds new adapter to Alice.
   void addAdapter(AliceAdapter adapter) => adapter.injectCore(_aliceCore);
+
+  /// Tag the next [maxCalls] HTTP requests with [tag].
+  /// If [maxCalls] is not provided, tags all subsequent calls until [clearTag] is called.
+  void tag({required String tag, int? maxCalls}) =>
+      _aliceCore.tagService.tag(tag: tag, maxCalls: maxCalls);
+
+  /// Clear the active tag.
+  void clearTag() => _aliceCore.tagService.clearTag();
 }